Refactor ContractsTests to use Java text blocks
- Migrated test data from external files to inline text blocks using addSourceLines.
- Deleted physical test files: CheckContractPositiveCases.java and CheckContractNegativeCases.java.
- Removed unnecessary @SuppressWarnings("deprecation") annotations.
- Part of the ongoing effort to modernize tests (#1425).
